refactor: split AgentToolRouter into modular handlers
