+++ app/src/main/java/cn/com/omnimind/bot/agent/tool/handlers/ToolHandler.kt
@@ -1,0 +1,24 @@
+package cn.com.omnimind.bot.agent.tool.handlers
+
+import cn.com.omnimind.bot.agent.AgentCallback
+import cn.com.omnimind.bot.agent.AgentToolExecutionHandle
+import cn.com.omnimind.bot.agent.AgentToolRegistry
+import cn.com.omnimind.bot.agent.ToolExecutionResult
+import kotlinx.serialization.json.JsonObject
+
+interface ToolHandler {
+    val toolNames: Set<String>
+
+    fun canHandle(toolName: String): Boolean = toolName in toolNames
+
+    suspend fun execute(
+        toolCall: cn.com.omnimind.baselib.llm.AssistantToolCall,
+        args: JsonObject,
+        runtimeDescriptor: AgentToolRegistry.RuntimeToolDescriptor,
+        env: cn.com.omnimind.bot.agent.AgentExecutionEnvironment,
+        callback: AgentCallback,
+        toolHandle: AgentToolExecutionHandle
+    ): ToolExecutionResult
+
+    suspend fun dispose() {}
+}
